보니까 값은 정상적으로 가는데 ui가 표시를 못하고 있다.

+++ Source/XRDefenceGame/Private/UI/MainInfoBoard.cpp
@@ -125,7 +125,7 @@ void AMainInfoBoard::UpdateInGameUI_Implementation(float TimeSecond, float Total
 	SetUITime(TimeSecond);
 }
 
-void AMainInfoBoard::SetFinalResultPanel_Multi_Implementation(int32 FirstNexusCount, int32 FirstNexusHealth, int32 FirstTimeLeft, int32 SecondNexusCount, int32 SecondNexusHealth, int32 SecondTimeLeft)
+void AMainInfoBoard::SetFinalResultPanel_Multi_Implementation(float FirstNexusCount, float FirstNexusHealth, float FirstTimeLeft, float SecondNexusCount, float SecondNexusHealth, float SecondTimeLeft)
 {
 	if (GEngine)
 	{
