+++ Source/XRDefenceGame/Public/Managet/XRDefenceGameInstance.h
@@ -34,13 +34,13 @@ public:
 	UPROPERTY(BlueprintReadWrite, VisibleAnywhere, Category = "Player")
 	FRotator DefencePlayerGamePlayRotation = FRotator(0.f, 180.f, 0.f);
 
-	int32 FirstNexusCount = -1;
-	int32 FirstNexusHealth = -1;
-	int32 FirstTimeLeft = -1;
+	float FirstNexusCount = -1;
+	float FirstNexusHealth = -1;
+	float FirstTimeLeft = -1;
 
-	int32 SecondNexusCount = -1;
-	int32 SecondNexusHealth = -1;
-	int32 SecondTimeLeft = -1;
+	float SecondNexusCount = -1;
+	float SecondNexusHealth = -1;
+	float SecondTimeLeft = -1;
 	
 	EObjectType ServerObjectType = EObjectType::EOT_Deffence;
 	EObjectType ClientObjectType = EObjectType::EOT_Offence;
